Change of dataset and changes in function names and arguments. Addition of new functions
We have changed the previous dataset:
   - Currently, we use the MTBLS4
Most of the functions names are now in snake_case. This has to be finished.
We have removed the 'lcms_' prefix for some of the datasets employed as examples. This has to be finished for all datasets for internal coherence.
New functions for post processing and data visulization are included now.
A filter function for mz is now included
Most of the functions have the examples comented. This has to be changed